# foundry/workiq_graph_openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Work IQ Graph Search (Auto FNOL Triage)",
    "description": "Public-endpoint (graph.microsoft.com) wrapper exposing SharePoint/OneDrive document search and Outlook mail search to the Foundry orchestrator agent. Backed by tools_workiq_graph.py. No VNet/Private Endpoint required.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://app-autofnol-workiq.azurewebsites.net" }
  ],
  "paths": {
    "/search-documents": {
      "post": {
        "operationId": "searchDocuments",
        "summary": "Search SharePoint/OneDrive documents (Work IQ document search equivalent).",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": { "type": "string", "description": "Natural-language search query, e.g. a storm/hurricane name or topic." },
                  "top": { "type": "integer", "default": 5 }
                },
                "required": ["query"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matching documents.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": { "type": "string" },
                      "webUrl": { "type": "string" },
                      "summary": { "type": "string" },
                      "lastModified": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search-mail": {
      "post": {
        "operationId": "searchMail",
        "summary": "Search Outlook mail in the Agent Identity's own mailbox (Work IQ mail search equivalent). Always searches the FNOL intake mailbox this agent runs as - no mailbox to specify.",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": { "type": "string", "description": "Natural-language search query, e.g. 'SIU routing change'." },
                  "top": { "type": "integer", "default": 5 }
                },
                "required": ["query"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matching emails.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "subject": { "type": "string" },
                      "from": { "type": "string" },
                      "receivedDateTime": { "type": "string" },
                      "bodyPreview": { "type": "string" },
                      "webLink": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/send-email": {
      "post": {
        "operationId": "sendEmail",
        "summary": "Send an email from the Agent Identity's own mailbox, e.g. to escalate a case to SIU or an adjuster team.",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to": { "type": "string", "description": "Comma-separated recipient email address(es)." },
                  "cc": { "type": "string", "description": "Optional comma-separated cc email address(es)." },
                  "subject": { "type": "string" },
                  "bodyText": { "type": "string", "description": "Plain-text email body." }
                },
                "required": ["to", "subject", "bodyText"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "to": { "type": "string" },
                    "subject": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}